V.1.0.0
Changelog 1.0.0:
- New:
- - Finished basic employee view
- - Time will be recorded now
- - Time can be seen on relevant views
- Fixes:
- - Employee will see only his own task groups
- - Time is now properly split between tasks

Changelog 0.2.1:
- New:
- - Added time display for main task group view
- - Added time display template for recycler view
- - Visuals finished - backend left

Changelog 0.2.0:
- New:
- - Base log-in system
- - Added default user 0 -Admin
- - Managers can now enter employee view
- - Finished base fragment layout and connections
- Fixes:
- - Only Admin can change Managers password
- - Only Admin can demote Managers

Changelog 0.1.5b:
- New:
- - Added recycler view sorting

Changelog 0.1.5:
- New:
- - Created password changing system
- - Added insight into which Task group the employee has access to
- - Manager can now manage the Employee - Task Group connections from Employee

Changelog 0.1.4b:
- Fixes:
- - Keeper no longer confuses User ID with GroupEmployee ID

Changelog 0.1.4:
- New:
- - Split files to make data management easier
- - Finished adding Employees to task groups
- - Created new DataClass containing Group worker and relevant id's
- - Keeper will now remember connection id's
- - Finished managing employees in task groups (In addition to adding, they can also be removed)
- Fixes:
- - Added colouration to templates to better see columns

Changelog 0.1.3b:
- New:
- - Updated naming convention (Location name is preceded by 3-4 letters od each word in source's name)
- - Fit template to Task Group's Employee's view

Changelog 0.1.3:
- New:
- - Added employee template
- - Permissions can now be changed by the manager
- - Added temp. password table for testing purposes
- - Manager needs to input password twice to add people
- - Added function to get last added employee's id
- - Manager's can now add employees and promote/demote them
- Fixes:
- - Manager can only remove and change perms of other employees, not of self or id = 0
- - Task status will now fit in one line

Changelog 0.1.2:
- New:
- - Keeper employed
- - Task View Adapter for Manager's Task Group Details added together with template
- - When entering Task Group's Task Details in managers side, it will now remember it's affiliation
- - Managers can now switch Task Status
- - Keeper will remember the user
- - Added relevant fields to database
- Fixes:
- - Added status to task in database
- - Recycler view templates now have clearer names, allowing them to be reused
- - Added Content Description to ImageButtons
- - Users are now called Employees

Changelog 0.1.1:
- New:
- - Added Task Group addition
- - Added Manager's view Task Group RV Adapter
- Fixes:
- - Fixed ID error for MVTGVDetailViewMenu
- - Fixed typo in database (missing ',')
- - Fixed Button-Bar

+++ app/src/main/java/com/example/whowherewhen/data/ExtendedTaskGroupData.kt
@@ -1,3 +1,3 @@
 package com.example.whowherewhen.data
 
-data class ExtendedTaskGroupData(val id: Int, var name: String, var time: Int)
+data class ExtendedTaskGroupData(val id: Int, var name: String, var time: Long)
